LPD-54772 Merge api and ApiHelper in one single file
