+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/services/FolderService.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {patch} from '../services/api';
+import {patch} from '../services/ApiHelper';
 import ApiHelper from './ApiHelper';
 
 export type TFolder = {
